java: Moving tests for RemoteWebDriver from server to client because there is no server specific in these tests
