+++ java/client/test/org/openqa/selenium/remote/RemoteWebDriverTest.java
@@ -15,27 +15,16 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium.remote.server;
+package org.openqa.selenium.remote;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import com.google.common.io.ByteStreams;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 
 import org.junit.Test;
 import org.openqa.selenium.Capabilities;
-import org.openqa.selenium.remote.CapabilityType;
-import org.openqa.selenium.remote.CommandExecutor;
-import org.openqa.selenium.remote.ErrorCodes;
-import org.openqa.selenium.remote.HttpCommandExecutor;
-import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.testing.JUnit4TestBase;
 
-import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 public class RemoteWebDriverTest extends JUnit4TestBase {
